refactor: CartViewModelFactory를 사용하여 ViewModel 인스턴스 생성
