+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModelFactory.kt
@@ -1,18 +1,15 @@
-package woowacourse.shopping.ui
+package woowacourse.shopping.ui.cart
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import woowacourse.shopping.data.CartRepository
 import woowacourse.shopping.data.CartRepositoryImpl
-import woowacourse.shopping.data.ProductRepository
-import woowacourse.shopping.data.ProductRepositoryImpl
 
 @Suppress("UNCHECKED_CAST")
-class MainViewModelFactory : ViewModelProvider.Factory {
+class CartViewModelFactory: ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        val productRepository: ProductRepository by lazy { ProductRepositoryImpl() }
         val cartRepository: CartRepository by lazy { CartRepositoryImpl() }
 
-        return MainViewModel(productRepository, cartRepository) as T
+        return CartViewModel(cartRepository) as T
     }
 }
